using express-async-handler package in backend

+++ backend/app/controllers/categories/deleteCategory.js
@@ -1,29 +1,23 @@
-// const mongoose = require('mongoose')
+const mongoose = require('mongoose')
+const asyncHandler = require('express-async-handler')
 const Category = require('../../models/category')
 
-const deleteCategory = async (req, res) => {
+const deleteCategory = asyncHandler(async (req, res) => {
   const { id } = req.params
 
-  // another way to check id
-  // if(!mongoose.isValidObjectId(id)) return res.status(400).json({ success: false, message: 'Category not found'})
-
-  try {
-    const category = await Category.findByIdAndRemove(id)
-
-    if (!category)
-      return res
-        .status(404)
-        .json({ success: false, message: 'Category not found.' })
-
-    res.json({ success: true, message: 'Category deleted successfully.' })
-  } catch (error) {
-    if (error.kind === 'ObjectId')
-      return res
-        .status(404)
-        .json({ success: false, message: 'Category not found.' })
-
-    res.status(500).json({ error, success: false })
+  if (!mongoose.isValidObjectId(id)) {
+    res.status(404)
+    throw new Error('Category not found.')
   }
-}
+
+  const category = await Category.findByIdAndRemove(id)
+
+  if (!category) {
+    res.status(404)
+    throw new Error('Category not found.')
+  }
+
+  res.json({ success: true, message: 'Category deleted successfully.' })
+})
 
 module.exports = deleteCategory
