+++ backend/app/controllers/users/deleteUser.js
@@ -1,25 +1,23 @@
+const asyncHandler = require('express-async-handler')
+const mongoose = require('mongoose')
 const User = require('../../models/user')
 
-const deleteUser = async (req, res) => {
+const deleteUser = asyncHandler(async (req, res) => {
   const { id } = req.params
 
-  try {
-    const user = await User.findByIdAndRemove(id)
-
-    if (!user)
-      return res
-        .status(404)
-        .json({ success: false, message: 'User not found.' })
-
-    res.json({ success: true, message: 'User deleted successfully.' })
-  } catch (error) {
-    if (error.kind === 'ObjectId')
-      return res
-        .status(404)
-        .json({ success: false, message: 'User not found.' })
-
-    res.status(500).json({ error, success: false })
+  if (!mongoose.isValidObjectId(id)) {
+    res.status(404)
+    throw new Error('User not found.')
   }
-}
+
+  const user = await User.findByIdAndRemove(id)
+
+  if (!user) {
+    res.status(404)
+    throw new Error('User not found.')
+  }
+
+  res.json({ success: true, message: 'User deleted successfully.' })
+})
 
 module.exports = deleteUser
